commit 20180508 VR and so on

+++ Assets/Scripts/CustomerScripts/PlayerBehaviorText_Fair2Ver.cs
@@ -109,7 +109,7 @@ public class PlayerBehaviorText_Fair2Ver : MonoBehaviour
 
 
         playerControllerObj = GameObject.Find("PlayerController");
-
+        
     }
 
 
@@ -164,7 +164,8 @@ public class PlayerBehaviorText_Fair2Ver : MonoBehaviour
             if (lastBehav != whichBehavior)
             {
                 //Debug.Log("whichBehav was changed from " + lastBehav + " to " + whichBehavior);
-                playerBehavList.Add(whichBehavior.ToString());
+                //playerBehavList.Add(whichBehavior.ToString());
+                playerBehavList.Add(((int)whichBehavior).ToString());
             }
 
 
@@ -185,7 +186,7 @@ public class PlayerBehaviorText_Fair2Ver : MonoBehaviour
             // エリアごとにbehavArrayを区切ってplayerBehavListに
             // playerBehavListにつないでいく
             playerBehavList.Add(WAT);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //playerBehavList.Add(behavArray);
             //behavArray.Clear();
             //behavArray.Add(WAT);
@@ -195,89 +196,89 @@ public class PlayerBehaviorText_Fair2Ver : MonoBehaviour
         {
             //sw.WriteLine(TA);
             playerBehavList.Add(TA);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter TA, " + TA);
         }
         else if (collider.gameObject.tag == "Sec_SHA")
         {
             //sw.WriteLine(SHA);
             playerBehavList.Add(SHA);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter SHA, " + SHA);
         }
         else if (collider.gameObject.tag == "Sec_YA")
         {
             //sw.WriteLine(YA);
             playerBehavList.Add(YA);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter YA, " + YA);
         }
         else if (collider.gameObject.tag == "Sec_KAI")
         {
             //sw.WriteLine(KAI);
             playerBehavList.Add(KAI);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter KAI, " + KAI);
         }
         else if (collider.gameObject.tag == "Sec_WAN")
         {
             //sw.WriteLine(WAN);
             playerBehavList.Add(WAN);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter WAN, " + WAN);
         }
         else if (collider.gameObject.tag == "Sec_YO")
         {
             //sw.WriteLine(YO);
             playerBehavList.Add(YO);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter YO, " + YO);
         }
         else if (collider.gameObject.tag == "Sec_KI")
         {
             //sw.WriteLine(KI);
             playerBehavList.Add(KI);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter KI, " + KI);
         }
         else if (collider.gameObject.tag == "Sec_KAK")
         {
             //sw.WriteLine(KAK);
             playerBehavList.Add(KAK);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter KAK, " + KAK);
         }
         else if (collider.gameObject.tag == "Sec_RI")
         {
             //sw.WriteLine(RI);
             playerBehavList.Add(RI);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
             //Debug.Log("Enter RI");
         }
         else if (collider.gameObject.tag == "Sec_FAN_1")
         {
             playerBehavList.Add(FAN_1);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
         }
         else if (collider.gameObject.tag == "Sec_FAN_2")
         {
             playerBehavList.Add(FAN_2);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
         }
         else if (collider.gameObject.tag == "Sec_F_STA")
         {
             playerBehavList.Add(STA_F);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
         }
         else if (collider.gameObject.tag == "Sec_S_STA")
         {
             playerBehavList.Add(STA_S);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
         }
         else if (collider.gameObject.tag == "Sec_COR")
         {
             playerBehavList.Add(COR);
-            playerBehavList.Add(WhichBehavior.WALK.ToString());
+            playerBehavList.Add(((int)WhichBehavior.WALK).ToString());
         }
     }
 }
